Renamed Destroyable to Disposable
