+++ storage/storage/src/main/java/one/microstream/storage/types/StorageChannel.java
@@ -41,12 +41,13 @@ import one.microstream.persistence.types.PersistenceIdSet;
 import one.microstream.persistence.types.Unpersistable;
 import one.microstream.storage.exceptions.StorageExceptionConsistency;
 import one.microstream.time.XTime;
+import one.microstream.typing.Disposable;
 import one.microstream.typing.KeyValue;
 import one.microstream.util.BufferSizeProviderIncremental;
 import one.microstream.util.logging.Logging;
 
 
-public interface StorageChannel extends Runnable, StorageChannelResetablePart, StorageActivePart, Destroyable
+public interface StorageChannel extends Runnable, StorageChannelResetablePart, StorageActivePart, Disposable
 {
 	public StorageTypeDictionary typeDictionary();
 
@@ -539,7 +540,7 @@ public interface StorageChannel extends Runnable, StorageChannelResetablePart, S
 			{
 				try
 				{
-					this.destroy();
+					this.dispose();
 				}
 				catch(final Throwable t1)
 				{
@@ -795,10 +796,10 @@ public interface StorageChannel extends Runnable, StorageChannelResetablePart, S
 		}
 
 		@Override
-		public final void destroy()
+		public final void dispose()
 		{
 			this.entityCache.reset();
-			this.fileManager.destroy();
+			this.fileManager.dispose();
 		}
 	}
